actualiza

+++ bin/gc.sh
@@ -6,6 +6,12 @@ if (test "$?" = "0") then {
 	echo "Gemfile inlcuye un sivel2_gen cableado al sistema de archivos"
 	exit 1;
 } fi;
+grep "^ *gem *.sivel2_sjr. *, *path:" Gemfile > /dev/null 2> /dev/null
+if (test "$?" = "0") then {
+	echo "Gemfile inlcuye un sivel2_sjr cableado al sistema de archivos"
+	exit 1;
+} fi;
+
 
 RAILS_ENV=test rake db:drop db:setup sivel2:indices
 if (test "$?" != "0") then {
